# optimize Lsh and Rsh