`Benchmark_Lsh`
20000000	        98.2 ns/op
300000000	         4.69 ns/op compared to 7.27 ns/op
PASS

`Benchmark_Rsh`
20000000	        81.5 ns/op
300000000	         4.70 ns/op compared to 7.22 ns/op
PASS