@kurkomisi 2018-03-05
`Lsh` 4.57 ns/op, `Rsh` 4.61 ns/op