@kurkomisi 2018-03-05
`Lsh` 4.53 ns/op, `Rsh` 4.48 ns/op